requirements.txt
ansible==2.9.0
ldap3==2.6.1
python-logstash==0.4.6
configparser==4.0.2
